+++ R/MiscFuns.R
@@ -11477,6 +11477,11 @@ initialize_startup_msg = function(startup_msg){
         return(FALSE)
     }
 
+    if(is_Rmarkdown()){
+        # Never in Rmarkdown: too ugly
+        return(FALSE)
+    }
+
     if(is.null(find_project_path())){
         return(FALSE)
     }