+++ R/onLoad.R
@@ -1,7 +1,7 @@
 
 
 
-.onLoad <- function(libname, pkgname){
+.onLoad = function(libname, pkgname){
 	# setting some options